minimongo: Find paused flag in the right place
The `paused` flag is stored on LocalCollection, but two places in
minimongo looked for it on the `query` object (which represents an
active observeChanges call).  In both cases, the bug had no correctness
impact but could have a performance impact.

Bug 1:

`_recomputeResults` (used to calculate changes to skip and limit
queries) tried to avoid calculating the diff if the collection had been
paused (by pauseObservers as part of latency compensation), but it
looked for the `paused` flag in the wrong place and always ran the diff.

This didn't have an effect on correctness (because the wrapped callbacks
on `query` are no-ops when `paused` is set) but did waste time on
unnecessary diffs.

Bug 2:

In `update`, we tried to avoid saving original results for skip/limit
queries if the collection was paused, because we don't actually run the
diff for paused queries.  (Well, except for the fact that Bug 1 made us
run the diffs anyway...)  But since we checked `query.paused` instead of
`self.paused`, we wasted time cloning the results even if we were
paused.